WIP - Moving more dependencies to PMFRGSolve
(e.g. DiffEqCallbacks, SciMLBase)
