+++ PMFRGCore.jl/Project.toml
@@ -4,14 +4,12 @@ authors = ["Nils Niggemann <nils.niggemann@fu-berlin.de>"]
 version = "2.2.1"
 
 [deps]
-DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
 FixedPoint = "3325f569-5a18-4e7d-8356-246b69339eea"
 H5Zblosc = "c8ec2601-a99c-407f-b158-e79c03c2f5f7"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
-SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SpinFRGLattices = "053fb7f8-31b7-4f68-aef8-4452b415dd8c"
 StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
