+++ PMFRGCore.jl/src/FileIO.jl
@@ -33,23 +33,6 @@ function readLam(Filename::String)
     return Lam
 end
 
-function readObservables(Filename::String, ObsType = Observables)
-    t = h5read(Filename, "Observables/Lambda")
-    Fields = fieldnames(ObsType)
-    obsTuple = Tuple(h5read(Filename, "Observables/$f") for f in Fields)
-    saveval = ObsType[]
-    saved_values = SavedValues(eltype(t), ObsType)
-    # return obsTuple
-    for i in eachindex(t)
-        Currentval(x) = selectdim(x, length(size(x)), i) |> Array
-        # return Currentval.(obsTuple)
-        push!(saveval, ObsType(Currentval.(obsTuple)...))
-    end
-    append!(saved_values.t, t)
-    append!(saved_values.saveval, saveval)
-    return saved_values
-end
-
 function readGeometry(Filename::String, GeometryGenerator::Function; Group = "Geometry")
     NLen = h5read(Filename, "$Group/NLen")
     couplings = h5read(Filename, "$Group/couplings")
@@ -125,21 +108,6 @@ end
 function setupDirectory(::Nothing, args...; kwargs...)
     return nothing
 end
-
-function saveCurrentState(
-    DirPath::String,
-    State::AbstractArray,
-    saved_Values::DiffEqCallbacks.SavedValues,
-    Lam::Real,
-    Par::PMFRGParams,
-)
-    Filename = joinpath(DirPath, "CurrentState.h5")
-    saveState(Filename, State, Lam, "w")
-    saveParams(Filename, Par)
-    saveObs(Filename, saved_Values, "Observables")
-    Filename
-end
-saveCurrentState(::Nothing, args...) = nothing
 
 """Rename CurrentState to FinalState as indicator that Job is finished"""
 function SetCompletionCheckmark(DirPath::String)
@@ -275,114 +243,6 @@ function getFileParams(Filename, Geometry, Par::Nothing; kwargs...)
     return readParams(Filename, Geometry; Lam_max = Lam, kwargs...)
 end
 
-"""Saves Observables"""
-function saveObs(
-    Filename::String,
-    saved_values::DiffEqCallbacks.SavedValues,
-    Group::String = "",
-)
-    ObsArr = StructArray(saved_values.saveval)
-    saveObs(Filename, ObsArr, Group)
-    h5write(Filename, joinGroup(Group, "Lambda"), saved_values.t)
-end
-
-function saveObs(Filename::String, Obs::StructArray{ObsType}, Group::String) where {ObsType}
-    Fields = fieldnames(ObsType)
-    for F in Fields
-        arr = convertToArray(getproperty(Obs, F))
-        h5write(Filename, joinGroup(Group, string(F)), arr)
-    end
-end
-
-
-function convertToArray(
-    VecOfArray::AbstractVector{VT},
-) where {N,VT<:AbstractArray{T,N} where {T}}
-    cat(VecOfArray..., dims = N + 1)
-end
-
-function saveMainOutput(Filename::String, saved_values, Group::String)
-    mkpath(dirname(Filename))
-    println("Saving Main output to ", abspath(Filename))
-    saveObs(Filename, saved_values, Group)
-end
-saveMainOutput(::Nothing, args...) = nothing
-
-
-function setCheckpoint(Directory::String, State, saved_values, Lam, Par, checkPointList)
-    saveCurrentState(Directory, State, saved_values, Lam, Par)
-    if !isempty(checkPointList)
-        if Lam < last(checkPointList)
-            Checkpoint = pop!(checkPointList)
-            CheckpointFile = UniqueFileName("$Directory/$(strd(Lam)).h5")
-            println("\nsaving Checkpoint Lam ≤ $Checkpoint at ", Lam)
-            println("in file ", CheckpointFile)
-            println("")
-            mv(joinpath(Directory, "CurrentState.h5"), CheckpointFile)
-        end
-    end
-end
-
-function setCheckpoint(Directory::Nothing, State, saved_values, Lam, Par, checkPointList)
-    return
-end
-
-function saveExtraFields(
-    Filename::String,
-    State,
-    Lambda::Real,
-    Par::PMFRGParams,
-    Group::String,
-)
-    (; T, N) = Par.NumericalParams
-    (; Name, NUnique, Npairs, NLen) = Par.System
-    Chi_nu = getChi(State, Lambda, Par, N)
-    h5write(Filename, "$Group/Name", Name)
-    h5write(Filename, "$Group/Npairs", Npairs)
-    h5write(Filename, "$Group/T", T)
-    h5write(Filename, "$Group/N", N)
-    h5write(Filename, "$Group/NUnique", NUnique)
-    h5write(Filename, "$Group/NLen", NLen)
-    h5write(Filename, "$Group/Chi_nu", Chi_nu)
-end
-
-saveMainOutput(
-    Filename::String,
-    Solution::ODESolution,
-    saved_values::DiffEqCallbacks.SavedValues,
-    Par::PMFRGParams,
-    Group::String,
-) = saveMainOutput(Filename, Solution.u[end], saved_values, saved_values.t[end], Par, Group)
-
-function saveMainOutput(
-    Filename::String,
-    State,
-    saved_values,
-    Lambda::Real,
-    Par::PMFRGParams,
-    Group::String,
-)
-    function save(name)
-        saveMainOutput(name, saved_values, Group)
-        saveExtraFields(name, State, Lambda, Par, Group)
-    end
-    try
-        save(Filename)
-    catch e
-        newName = UniqueFileName(Filename)
-        @warn "Writing to $Filename errored with exception $(string(e))! Writing to $newName instead."
-        save(newName)
-    end
-    # saveParams(Filename,Par,Group)
-end
-
-saveMainOutput(
-    Filename::String,
-    Solution::ODESolution,
-    saved_values::DiffEqCallbacks.SavedValues,
-    Par::PMFRGParams,
-    Group::Nothing,
-) = saveMainOutput(Filename, Solution, saved_values, Par, string(Par.NumericalParams.T))
 
 function getFilesFromSubDirs(Folder::String)
     allpaths = collect(walkdir(Folder))
