+++ PMFRGCore.jl/src/MultiLoop/MultiLoopPMFRG.jl
@@ -5,7 +5,6 @@ includes MultiLoop corrections to PMFRG.
 include("MultiLoopTypes.jl")
 include("Parquet.jl")
 
-export SolveParquet
 generateFileName(Par::MultiLoopParams, arg::String = "") =
     _generateFileName(Par, "_l$(Par.l)" * arg)
 generateFileName(Par::ParquetParams, arg::String = "") = _generateFileName(Par, "_p" * arg)
