+++ PMFRGCore.jl/src/SolverUtils.jl
@@ -55,85 +55,6 @@ t_to_Lam(t) = exp(t)
 """Converts physical cutoff Λ to t (integrator step). Inverse of t_to_Lam."""
 Lam_to_t(t) = log(t)
 
-function get_output_callback(
-    CheckPointSteps,
-    CheckpointDirectory,
-    saved_values,
-    Par,
-    VertexCheckpoints,
-)::SciMLBase.DiscreteCallback
-    # count number of outputs = number of steps.
-    # CheckPointSteps gives the intervals in which checkpoints should be saved.
-    i = 0
-    function bareOutput(State, t, integrator)
-        @timeit_debug "bareOutput" begin
-            Lam = t_to_Lam(t)
-            i += 1
-            i % CheckPointSteps == 0 && setCheckpoint(
-                CheckpointDirectory,
-                State,
-                saved_values,
-                Lam,
-                Par,
-                VertexCheckpoints,
-            )
-        end
-    end
-
-    function verboseOutput(State, t, integrator)
-        @timeit_debug "verboseOutput" begin
-            Lam = t_to_Lam(t)
-            println("Time taken for output saving: ")
-            bareOutput(State, t, integrator)
-            println("")
-            writeOutput(State, saved_values, Lam, Par)
-        end
-    end
-
-    function getOutputfunction(MinimalOutput)
-        if MinimalOutput
-            return bareOutput
-        else
-            return verboseOutput
-        end
-    end
-    output_func = getOutputfunction(Par.Options.MinimalOutput)
-    outputCB = FunctionCallingCallback(output_func, tdir = -1, func_start = false)
-    return outputCB
-end
-
-
-function get_saving_callback(
-    ObservableType,
-    Par,
-    ObsSaveat,
-    Lam_min,
-    Lam_max,
-    saved_values,
-)::SciMLBase.DiscreteCallback
-    #get Default for lambda range for observables
-    # ObsSaveat = getLambdaMesh(ObsSaveat,Lam_min,Lam_max)
-    save_func(State, t, integrator) =
-        getObservables(ObservableType, State, t_to_Lam(t), Par)
-    ObsSaveat = gettMesh(ObsSaveat, Lam_min, Lam_max)
-    saveCB = SavingCallback(
-        save_func,
-        saved_values,
-        save_everystep = false,
-        saveat = ObsSaveat,
-        tdir = -1,
-    )
-    return saveCB
-end
-
-function get_problem(Lam_max, Lam_min, State, setup, Deriv!)::ODEProblem
-    t0 = Lam_to_t(Lam_max)
-    tend = get_t_min(Lam_min)
-    Deriv_subst! = generateSubstituteDeriv(Deriv!)
-    problem = ODEProblem(Deriv_subst!, State, (t0, tend), setup)
-    return problem
-end
-
 
 function generateSubstituteDeriv(getDeriv!::Function)
