+++ PMFRGCore.jl/test/UnitTests/UnitTests.jl
@@ -36,7 +36,7 @@ include("DimerTest.jl")
 
 include("TypeStability.jl")
 
-include("IOTests.jl")
+# include("IOTests.jl") # DEBUG
 include("ParquetTest.jl")
 
 function testOneLoop(Obsacc = 1e-14)
