+++ PMFRGSolve.jl/Project.toml
@@ -8,7 +8,9 @@ DiffEqCallbacks = "459566f4-90b8-5000-8ac3-15dfb0a30def"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 OrdinaryDiffEqLowOrderRK = "1344f307-1e59-4825-a18e-ace9aa3fa4c6"
 PMFRGCore = "658c49e6-d26a-4ae0-a540-09812f09d5ba"
+SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SpinFRGLattices = "053fb7f8-31b7-4f68-aef8-4452b415dd8c"
+StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 
 [weakdeps]
