+++ PMFRGSolve.jl/src/PMFRGSolve.jl
@@ -1,13 +1,16 @@
 module PMFRGSolve
 using OrdinaryDiffEqLowOrderRK, DiffEqCallbacks
-using PMFRGCore, SpinFRGLattices, TimerOutputs
+using PMFRGCore, SpinFRGLattices, TimerOutputs, StructArrays
 
-export BS3, Vern7, DP5
+export BS3, Vern7, DP5, readObservables
 
 include("FileIO.jl")
 include("Solver.jl")
 include("TwoLoop/TwoLoopPMFRG.jl")
 
-export SolveFRG, SolveFRG_Checkpoint
+export SolveFRG, SolveFRG_Checkpoint, saveCurrentState, saveMainOutput
+
+include("MultiLoop/Parquet.jl")
+export SolveParquet
 
 end # module PMFRGSolve
